requirements.txt
aiohttp==3.9.5
aiohttp==3.9.5
aiosignal==1.3.1
async-timeout==4.0.3
atomicwrites==1.4.1
attrs==23.2.0
autopep8==2.1.0
backports.tarfile==1.1.1
base58==2.1.1
bitarray==2.9.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
ckzg==1.0.1
click==8.1.7
colorama==0.4.6
cytoolz==0.12.1
dateparser==1.0.0
distlib==0.3.8
docutils==0.21.2
dydx-v3-python==2.1.0
ecdsa==0.19.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
filelock==3.13.4
flake8==7.0.0
frozenlist==1.4.1
hexbytes==0.3.1
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
ipfshttpclient==0.8.0a2
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
keyring==25.1.0
loguru==0.7.2
lru-dict==1.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
mpmath==1.0.0
multiaddr==0.0.9
multidict==6.0.5
mypy-extensions==1.0.0
netaddr==1.2.1
nh3==0.2.17
numpy==1.26.4
packaging==24.0
pandas==2.2.2
parsimonious==0.8.1
pathspec==0.12.1
patsy==0.5.6
pkginfo==1.10.0
platformdirs==4.2.1
pluggy==1.5.0
protobuf==3.19.5
py==1.11.0
pycodestyle==2.11.1
pycryptodome==3.20.0
pyflakes==3.2.0
Pygments==2.17.2
PyJWT==2.8.0
pyproject-api==1.6.1
pyrsistent==0.20.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
pytz-deprecation-shim==0.1.0.post0
pyunormalize==15.1.0
pywin32==306
pywin32-ctypes==0.2.2
readme_renderer==43.0
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-mock==1.12.1
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rlp==2.0.1
rpds-py==0.18.0
scipy==1.13.0
six==1.14.0
statsmodels==0.14.2
sympy==1.6
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tox==4.14.2
twine==4.0.2
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
varint==1.0.2
virtualenv==20.26.0
wcwidth==0.2.13
web3==5.31.4
web3_account==1.0.1
websockets==9.1
win32-setctime==1.1.0
yarl==1.9.4
zipp==3.18.1